Separate model test from view test
