+++ aidants_connect_web/tests/test_models.py
@@ -1,0 +1,26 @@
+from django.test import TestCase
+from aidants_connect_web.models import Connection
+
+
+class ConnectionModelTest(TestCase):
+    def test_saving_and_retrieving_connexion(self):
+        first_connexion = Connection()
+        first_connexion.state = "aZeRtY"
+        first_connexion.code = "ert"
+        first_connexion.nonce = "varg"
+        first_connexion.save()
+
+        second_connexion = Connection()
+        second_connexion.state = "QsDfG"
+        second_connexion.save()
+
+        saved_items = Connection.objects.all()
+        self.assertEqual(saved_items.count(), 2)
+
+        first_saved_item = saved_items[0]
+        second_saved_item = saved_items[1]
+
+        self.assertEqual(first_saved_item.state, "aZeRtY")
+        self.assertEqual(first_saved_item.nonce, "varg")
+        self.assertEqual(second_saved_item.state, "QsDfG")
+
